+++ packages/blog-core/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./schemas/index.js"
 export * from "./seo/index.js"
 export * from "./utils/index.js"
+export * from "./pages/index.js"